+++ packages/iov-jsonrpc/src/id.ts
@@ -1,0 +1,5 @@
+let counter = 0;
+
+export function makeId(): number {
+  return (counter += 1);
+}