+++ packages/iov-jsonrpc/src/jsoncompatibledictionary.spec.ts
@@ -1,3 +1,4 @@
+import { makeId } from "./id";
 import {
   isJsonCompatibleArray,
   isJsonCompatibleDictionary,