+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.13.7
 
+- @iov/jsonrpc: Add `makeId`.
 - @iov/socket: Add 10 second timeout to `SocketWrapper`. Value can be configured
   using an optional constructor argument.
 - @iov/socket: Add 10 second timeout to `StreamingSocket`. Value can be